Move PartialLikelihoods to their own file
